Print the size of various types

+++ 0x00-hello_world/6-size.c
@@ -1,18 +1,18 @@
 #include <stdio.h>
 /**
- * main: Entry point of program
+ * main - is the entry point for the program.
  *
- * return: 0 if no error
+ * return: zero if no error, non-zero value if errors.
  */
 
 int main(void)
 {
 	
-	printf("Size of a char: %d byte(s)\n", (unsigned long)sizeof(char));
-	printf("Size of an int: %d byte(s)\n", (unsigned long)sizeof(int));
-	printf("Size of a long int: %d byte(s)\n", (unsigned long)sizeof(long int));
-	printf("Size of a long long int: %d byte(s)\n", (unsigned long)sizeof(long long int));
-	printf("Size of a float: %d byte(s)\n", (unsigned long)sizeof(float));
+	printf("Size of a char: %d byte(s)\n", sizeof(char));
+	printf("Size of an int: %d byte(s)\n", sizeof(int));
+	printf("Size of a long int: %d byte(s)\n", sizeof(long int));
+	printf("Size of a long long int: %d byte(s)\n", sizeof(long long int));
+	printf("Size of a float: %d byte(s)\n", sizeof(float));
 	return (0);
 
 }
